Added unit tests for the Model getters/setters.

+++ src/ch/epfl/leb/sass/ijplugin/Model.java
@@ -152,7 +152,7 @@ public class Model implements Serializable {
     public String getBackgroundCurrentSelection() {
         return backgroundCurrentSelection;
     }
-    public double getBackgroundUniformSignal() {
+    public float getBackgroundUniformSignal() {
         return backgroundUniformSignal;
     }
     public String getBackgroundTifFile() {
